+++ packages/playground/src/Form/FieldColorDemo.tsx
@@ -1,19 +1,14 @@
 /*
-
  MIT License
-
  Copyright (c) 2019 Looker Data Sciences, Inc.
-
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
  in the Software without restriction, including without limitation the rights
  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  copies of the Software, and to permit persons to whom the Software is
  furnished to do so, subject to the following conditions:
-
  The above copyright notice and this permission notice shall be included in all
  copies or substantial portions of the Software.
-
  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -21,57 +16,30 @@
  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  SOFTWARE.
-
  */
-import omit from 'lodash/omit'
-import styled from 'styled-components'
-import { CompatibleHTMLProps, reset } from '@looker/design-tokens'
-import {
-  border,
-  BorderProps,
-  height,
-  HeightProps,
-  width,
-  WidthProps,
-} from 'styled-system'
-import {
-  CustomizableInputTextAttributes,
-  inputTextDefaults,
-  inputTextDisabled,
-  inputTextHover,
-} from '../../../Inputs/InputText'
 
-export interface SwatchProps
-  extends BorderProps,
-    WidthProps,
-    HeightProps,
-    CompatibleHTMLProps<HTMLDivElement> {
-  /**
-   * The background color to display on the swatch.
-   */
-  color?: string
-}
+import React from 'react'
+import { Box, FieldColor, FieldSelect } from '@looker/components'
 
-export const Swatch = styled.div<SwatchProps>`
-  ${reset}
-  ${width}
-  ${height}
-  ${border}
+export function FieldColorDemo() {
+  const [color, setColor] = React.useState('red')
 
-  &:hover {
-    ${inputTextHover}
+  function handleChange(value: string) {
+    setColor(value)
   }
 
-  ${props => (props.disabled ? inputTextDisabled : '')}
+  function handleColorChange(e: React.FormEvent<HTMLInputElement>) {
+    setColor(e.currentTarget.value)
+  }
 
-  background-color: ${props => props.color};
-  margin-top: auto;
-  flex-shrink: 0;
-`
-
-Swatch.defaultProps = {
-  ...inputTextDefaults,
-  ...omit(CustomizableInputTextAttributes, 'fontSize'),
-  color: 'white',
-  width: CustomizableInputTextAttributes.height,
+  return (
+    <Box p="large" width={300}>
+      <FieldSelect
+        options={[{ value: 'green' }, { value: 'purple' }, { value: 'red' }]}
+        defaultValue="red"
+        onChange={handleChange}
+      />
+      <FieldColor value={color} onChange={handleColorChange} />
+    </Box>
+  )
 }